requirements.txt
pillow
transformers
colpali-engine[train]
accelerate
bitsandbytes
flash-attn
datasets
pynvml
wandb
mteb